Fix fromStream to handle already completed streams
In the implementation of `StreamListener` for `fromStream`, the `complete()` handler (for a completed observable stream) [calls `this.subscription.unsubscribe()`][unsub].

Unfortunately, `this.subscription` is initialized with the line `this.subscription = observable.subscribe(this)` in the constructor. This requires first evaluating `observable.subscribe(this)`, which calls the `complete()` handler if an already-completed observable was passed to it. As such, for already-completed observables, `fromStream` fails, trying to call `.unsubscribe()` on the `this.subscription` before it is initialized.

[unsub]: https://github.com/mobxjs/mobx-utils/blob/89f3da595cf1c22697ba701d101e6ad31c001f64/src/observable-stream.ts#L87
[sub]: https://github.com/mobxjs/mobx-utils/blob/89f3da595cf1c22697ba701d101e6ad31c001f64/src/observable-stream.ts#L74