Support IComputedOptions in createTransformer
In the same vain as #215, it would be great to use a custom equals
function with createTransformer, to avoid bubbling bubbling recomputes
of transformers that return arrays.

This commit passes the options through to the internal `computed`.  The
name option is not passed through, since that can be changed with
the `debugNameGenerator` option.